+++ Common/src/main/java/com/tristankechlo/whatdidijustkill/network/ClientBoundEntityKilledPacket.java
@@ -1,8 +1,6 @@
 package com.tristankechlo.whatdidijustkill.network;
 
 import com.tristankechlo.whatdidijustkill.client.ToastHandler;
-import com.tristankechlo.whatdidijustkill.config.WhatDidIJustKillConfig;
-import com.tristankechlo.whatdidijustkill.config.types.EntityOptions;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
@@ -28,13 +26,6 @@ public record ClientBoundEntityKilledPacket(Component entityName, ResourceLocati
 
     /* handle the packet; forge, fabric and neoforge */
     public static void handle(ClientBoundEntityKilledPacket packet) {
-        if (WhatDidIJustKillConfig.get().entity().showToast() == EntityOptions.ShowToastOption.NONE) {
-            return;
-        }
-
-        if (WhatDidIJustKillConfig.get().entity().isEntityExcluded(packet.entityType)) {
-            return;
-        }
         ToastHandler.showToastEntity(packet.entityName, packet.entityType, packet.distance, packet.hasSpecialName);
     }