+++ Changelog.md
@@ -7,6 +7,9 @@
     - style: `ADVANVEMENT` => dark background
     - style: `RECIPE` => light background, round corners
     - style: `TUTORIAL` => light background, no round corners
+- add key binding to toggle toast visibility
+    - default is CTRL+V
+    - configurable through the key bindings
 
 ### Version 1.20.4 - 1.0.0